Troubleshot missing 0xE5 token
The SD card was commanded to disable itself whenever it didn't
receive an 0xE5 token after a single block write. Although the
SD card would write several iterations' worth of data, it would
usually be unable to write as many iterations as desired.

I tried looking at and printing the response tokens as described
in the SD card specsheet, Section 7.2.4 and Section 7.3.3.1. I
was unable to find a solution. This commit is being done to
document the changes I made while troubleshooting.